+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/hardware/Volume.kt
@@ -1,0 +1,3 @@
+package com.atlassian.performance.tools.awsinfrastructure.api.hardware
+
+class Volume(val size: Int)